fix CR comments
