+++ mlrun/runtimes/sparkjob/spark3job.py
@@ -22,6 +22,7 @@ import mlrun.runtimes.pod
 from mlrun.config import config
 
 from ...execution import MLClientCtx
+from ...k8s_utils import validate_node_selectors
 from ...model import RunObject
 from ...utils import update_in, verify_field_regex
 from ..kubejob import KubejobRuntime
@@ -505,7 +506,7 @@ class Spark3Runtime(KubejobRuntime):
             raise NotImplementedError(
                 "Setting node name is not supported for spark runtime"
             )
-        mlrun.utils.validate_node_selectors(node_selector, raise_on_error=False)
+        validate_node_selectors(node_selector, raise_on_error=False)
         self.with_driver_node_selection(node_name, node_selector, affinity, tolerations)
         self.with_executor_node_selection(
             node_name, node_selector, affinity, tolerations
@@ -538,7 +539,7 @@ class Spark3Runtime(KubejobRuntime):
         if affinity is not None:
             self.spec.driver_affinity = affinity
         if node_selector is not None:
-            mlrun.utils.validate_node_selectors(node_selector, raise_on_error=False)
+            validate_node_selectors(node_selector, raise_on_error=False)
             self.spec.driver_node_selector = node_selector
         if tolerations is not None:
             self.spec.driver_tolerations = tolerations
@@ -570,7 +571,7 @@ class Spark3Runtime(KubejobRuntime):
         if affinity is not None:
             self.spec.executor_affinity = affinity
         if node_selector is not None:
-            mlrun.utils.validate_node_selectors(node_selector, raise_on_error=False)
+            validate_node_selectors(node_selector, raise_on_error=False)
             self.spec.executor_node_selector = node_selector
         if tolerations is not None:
             self.spec.executor_tolerations = tolerations
