+++ tests/test_k8s_utils.py
@@ -83,8 +83,62 @@ def test_sanitize_label_value(value: str, expected: str):
         ("invalid-key_", pytest.raises(mlrun.errors.MLRunInvalidArgumentError)),
         # trailing dot
         ("invalid-key.", pytest.raises(mlrun.errors.MLRunInvalidArgumentError)),
+        ("invalid-key/", pytest.raises(mlrun.errors.MLRunInvalidArgumentError)),
     ],
 )
 def test_verify_label_key(label_key, exception):
     with exception:
-        mlrun.k8s_utils.verify_label_key(label_key)
+        mlrun.k8s_utils.verify_label_key(label_key, allow_k8s_prefix=False)
+
+
+@pytest.mark.parametrize(
+    "node_selectors, expected",
+    [
+        # Valid cases
+        # node selectors that should pass validation
+        ({"kubernetes.io/arch": "amd64", "tier": "backend"}, does_not_raise()),
+        ({"datacenter/region": "us-west", "role": "worker"}, does_not_raise()),
+        ({"team/department": "engineering", "project": "ml-models"}, does_not_raise()),
+        ({"test": "", "kubernetes.io/os": "linux"}, does_not_raise()),
+        # Invalid cases
+        # Invalid key with extra slashes
+        (
+            {"invalid/key/format/with/slash": "value"},
+            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
+        ),
+        # Invalid character '=' in key
+        (
+            {"key-with-invalid-characters=": "value"},
+            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
+        ),
+        # Key with trailing slash
+        (
+            {"key-with-dash/": "value"},
+            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
+        ),
+        # Invalid character '=' in value
+        (
+            {"key": "value_with_invalid_chars=a"},
+            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
+        ),
+        # Value too long
+        (
+            {
+                "key": "value_with_very_long_string_that_exceeds_the_maximum_length_limit_of_63_characters"
+            },
+            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
+        ),
+        # Invalid character '#' in value
+        (
+            {"key": "value_with_invalid_character#"},
+            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
+        ),
+        # Value starts with a character
+        ({"key": ".valid-val"}, pytest.raises(ValueError)),
+        # Prefix is too long
+        ({"a" * 254 + "/key": "value"}, pytest.raises(ValueError)),
+    ],
+)
+def test_validate_node_selectors(node_selectors, expected):
+    with expected:
+        mlrun.k8s_utils.validate_node_selectors(node_selectors)
