+++ tests/utils/test_helpers.py
@@ -1093,56 +1093,3 @@ def test_validate_single_def_handler_valid_handler(code):
         pytest.fail(
             "validate_single_def_handler raised MLRunInvalidArgumentError unexpectedly."
         )
-
-
-@pytest.mark.parametrize(
-    "node_selectors, expected",
-    [
-        # Valid cases
-        # node selectors that should pass validation
-        ({"kubernetes.io/arch": "amd64", "tier": "backend"}, does_not_raise()),
-        ({"datacenter/region": "us-west", "role": "worker"}, does_not_raise()),
-        ({"team/department": "engineering", "project": "ml-models"}, does_not_raise()),
-        ({"test": "", "kubernetes.io/os": "linux"}, does_not_raise()),
-        # Invalid cases
-        # Invalid key with extra slashes
-        (
-            {"invalid/key/format/with/slash": "value"},
-            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
-        ),
-        # Invalid character '=' in key
-        (
-            {"key-with-invalid-characters=": "value"},
-            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
-        ),
-        # Key with trailing slash
-        (
-            {"key-with-dash/": "value"},
-            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
-        ),
-        # Invalid character '=' in value
-        (
-            {"key": "value_with_invalid_chars=a"},
-            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
-        ),
-        # Value too long
-        (
-            {
-                "key": "value_with_very_long_string_that_exceeds_the_maximum_length_limit_of_63_characters"
-            },
-            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
-        ),
-        # Invalid character '#' in value
-        (
-            {"key": "value_with_invalid_character#"},
-            pytest.raises(mlrun.errors.MLRunInvalidArgumentError),
-        ),
-        # Value starts with a character
-        ({"key": ".valid-val"}, pytest.raises(ValueError)),
-        # Prefix is too long
-        ({"a" * 254 + "/key": "value"}, pytest.raises(ValueError)),
-    ],
-)
-def test_validate_node_selectors(node_selectors, expected):
-    with expected:
-        mlrun.utils.validate_node_selectors(node_selectors)
